requirements.txt
requests==2.31.0
streamlit==1.29.0
websocket-client==1.7.0
setuptools==68.2.0
numpy==1.26.2
Pillow==10.1.0
msrest==0.7.1
matplotlib==3.8.2